custom-elements/instantiating-custom-elements/custom-element-type-local-name-and-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is the local name of the custom element even if IS attribute provided</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If both types of custom element types are provided at the time of element's instantiation, the custom tag must win over the type extension">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    doc.register('x-foo-1');
    doc.register('x-foo-2');
    
    doc.body.innerHTML = '<x-foo-1 id="xfoo" is="x-foo-2"></x-foo-1>';
    
    var xfoo = doc.querySelector('#xfoo');
    
    assert_equals(xfoo.constructor.prototype.constructor.name, xfoo.localName, 
        'Wrong custom element local name');

    assert_equals(xfoo.constructor.prototype.constructor.name, 'x-foo-1', 
        'Wrong custom element type');
    
}, 'Custom element type is the local name of the custom element even if IS attribute provided.');


test(function() {
    var doc = newHTMLDocument();
        
    doc.register('x-foo-4');
    
    doc.body.innerHTML = '<x-foo-3 id="xfoo" is="x-foo-4"></x-foo-3>';
    
    var xfoo = doc.querySelector('#xfoo');
    
    assert_equals(xfoo.constructor.prototype.constructor.name, 'HTMLElement', 
        'Wrong custom element local name');

    doc.register('x-foo-3');
    
    assert_equals(xfoo.constructor.prototype.constructor.name, xfoo.localName, 
            'Wrong custom element local name');

    assert_equals(xfoo.constructor.prototype.constructor.name, 'x-foo-3', 
            'Wrong custom element type');
    
}, 'Custom element type is the local name of the custom element even if IS attribute provided.'
    + 'Custom element is unresolved at first');

</script>
</body>
</html>
